Fixed warnings from Xcode 5.1 due to armv64

+++ ImageCachingTest/ImageCachingTest/AFNetworking/BPAFNetworkingViewController.m
@@ -43,7 +43,7 @@
     
     BPTableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:kBPCellID];
     
-    cell.textLabel.text = [NSString stringWithFormat:@"%d %d", indexPath.section, indexPath.row];
+    cell.textLabel.text = [NSString stringWithFormat:@"%ld %ld", (long)indexPath.section, (long)indexPath.row];
     NSURL *url = [self imageUrlForIndexPath:indexPath];
     cell.imageUrl = url;
     cell.customImageView.image = nil;
